pyInterface: Move initialization of printing counter.
Initialize the printing counter and the printing classes in the printingUtils
submodule instead of at the beginning of every script. This seems to work,
however, somehow there seem to be troubles with the silencer class. As the
whole package is in the process of being rewritten, the details will be ironed
out once the new version is more or less stable.