pyInterface: convertTreeToEvt.py
Add new python script converting ROOTPWA formatted root files to ASCII evt
files. This is a first version that still has some problems. Currently the
charge is set to zero in in the evt file even if the particle does carry
charge. Some changes in the python wrapper for the particleDataTable need to
happen first. There is no exception handling yet. Only the momenta of the
particles are translated to the evt file for now.