pyInterface: add amplitudeBenchmark.py
add a script that runs a benchmark test if the result of the calcAmplitudes
script is consistent with the result of the script on a benchmark machine. Also
add a script that compares 2 amplitude files. The script will save 2 histograms
for the differences between the files, one for the real and one for the
imaginary parts.